requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
django==2.1.7
djangorestframework==3.9.2
dj_database_url==0.5.0
gunicorn==19.6.0
httpie==0.9.8
httplib2==0.9.2
MarkupSafe==1.0
netifaces==0.10.4
Pillow==5.1.0
psutil==5.4.2
pycurl==7.43.0.1
psycopg2==2.7.7
requests==2.18.4
urllib3==1.22
whitenoise==4.1.2
